scylla_setup: update the node exporter service name
In commit 95197a09c91eb719c252abed847b8c035f5f9600, the name
of the node-exporter service was changed to be product dependant
because it is now build as a package.
When this commit was brought to enterprise, scylla_setup became
invalid since it contains a reference to `scylla-node-exporter`
which is automatically becomes `scylla-enterprise-node-exporter`
because of the product name.
This commit adjusts `scylla_setup`.

Fixes #1620

Closes #1621