audit: enable some subset of auditing by default (for enterprise)
It makes sense to enable a subset of auditing for debug purposes,
the enabled subset should have minimal impact on performance. The
logging method is set to be table as it is the option that make
more sense for cloud users. It has the additional benefit of being
docker compatible out of the box even though our docker solutions
already  have been adopted with rsyslog package for allowing logging
to syslog from within docker.

Fixes #1097
Fixes #1524