encryption: Add end-of-file marker to CL options metadata
Fixes #1682

If we crash between a CL encryptions options file was created not finished writing/flushed,
we will on restart open the corresponding segments for replay, but when reading the encryption
options we get partial/empty info and revert to "defaults" (which is repl provider).

Need to explicitly mark and check that we've actually read a full options file, and only then
wrap a file opened for read.

If we don't have a full options file, the segment has not been used, and is not readable
anyway.

Closes #1705