ICS: Fix use-after-free when estimating pending tasks
After the commit c3b8757fa1c5aa, sstable_set::all() no longer guarantee that
the set will keep a reference to the returned list, so it's no longer
safe to iterate through the list without holding a reference to it.
Commit 05b07c71616 changed range-based loops which relied on the API,
but we missed this usage of sstable_set::all() in ICS, which explains
the use-after-free.

Fixes #1728.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20210427233412.95627-1-raphaelsc@scylladb.com>