encryption: Check for commitlog encryption opts existence
Fixes #1681

If something managed to crash the node between a commitlog segment
file and the encryption properties file were created, we could
end up trying to open a non-existent file -> exception.

Calling code does not really handle this, so be graceful and
simply don't wrap in this case. Reading will almost certainly
fail, but that is not our problem.

Closes #1763