scripts/pull_github_pr.sh: don't rely on alias "enterprise" for remote
Currently, the script scripts/pull_github_pr.sh assumes that the
user set up an alias "enterprise" for the remote scylla-enterprise.git.
While some people are used to that, it's by no means necessary, or even
the default (if you just "git clone" scylla-enterprise.git, the default
name for the remote is "origin", not "enterprise").

A simple solution is not to assume any remote name at all, and just
use the URL directly: git@github.com:scylladb/scylla-enterprise.git.
It's not as if the user has any other option - the only place we can
find the pull requests is the official repository - it's not like any
other clone might have pull requests in them. Also the enterprise
repository requires authentication, so the git@ part isn't optional.

A more elaborate solution could be to run "git origin -v" and look
for scylla-enterprise.git, but I think that would be overkill and the
solution in this patch is simpler and good enough.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20210527145301.702503-1-nyh@scylladb.com>