auth/ldap_role_manager: Reconnect on query failure
Previously we called reconnect() only when the initial connection to
LDAP failed.  But the connection can fail at any time, leaving
ldap_role_manager in a permanently degraded state, even if the
connection failure was just temporary.

This patch fixes that by adding an ldap_reuser mechanism that shares a
connection while it's live and creates a new one when it fails.
Reconnecting is attempted only by the first fiber to notice a
connection error; others refrain from concurrent attempts.

Error reporting is improved: checks are done earlier, every error
results in an exceptional future, and the messages are more
descriptive.

A new test case exposed a bug in ldap_connection::write(), which was
fixed.

Fixes #1772.

Tests: unit (dev, debug)

Signed-off-by: Dejan Mircevski <dejan@scylladb.com>
Message-Id: <20210610143318.316450-1-dejan@scylladb.com>