Workload Prioritization: Add service levels creation robustness
We have a problem today with creating too many service levels,
too many is when there are more service levels than available
secheduling groups.
When it happens, the user will be confused by a constant message
about failure to create a scheduling group and what is worst is
that he can't understand from this what service levels are in the
system.
This change will prevent addition of a new service level if there
are no available scheduling groups. It does that in a very simple
way and does not take into account race conditions, taking those
into account is too complicated relative to the gain we get from
this and it will naturally be solved when we introduce raft.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>