Workload Prioritization: rename scheduling group for better stats
Up until now the scheduling groups attached to service levels were
named with a generic name. This convention made it nearly impossible
to determine which scheduling group is created for which service level.
Moreover, since we avoid deleting scheduling groups and just reuse them
this can be even more confusing since the metrics sometimes shows
scheduling groups for "dead" service level with no way to distinct
betwin the two.
This changes the convention to name the scheduling goroup after the
service level, this will make the metrics more clear and avoid confusion
in the future.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>