local_file_provider: Handle key lines w/o block/padding info
Fixes #1923

A key can have a descriptor string line "AES:128:<data>" iff user relies on
algo defaults. Must match this as well when reading file back, otherwise we
will of course fail reading anything encrypted with this key, and worse,
probably overwrite it since we now think it is missing from the file.

Closes #1928