encryption::kmip_host: Move and share some key defaults/validation between key/kmip
Fixes #1962

We now more or less support partial key specs. This can work or not work with KMIP
depending on the sever behaviour. In particular, we enforce block mode for keys if
not specified in openssl key creation. KMIP remote key lookup need to reflect this
better, otherwise we might get errors that don't reflect local behaviour.
Same with padding.

This moves moves some of the verification/default values from symmetric_key constructor
to shared code key/kmip_host::make_attributes (note that parsing needs to remain
"stupid" because it is also for comparisons).

Basically we need to actually fill in the block mode properly.

Closes #1963