auth,ldap: allow auto-creating users with assigned roles
This commit adds a way of auto-creating a user if it is already
granted roles in the LDAP role manager.
If the user tries to log in and its role was never created with
`CREATE USER`, but it already has roles assigned to it by the LDAP
role manager, the role manager will try to auto-create this user.
Before that, each user must have been created manually both in LDAP
and, separately, locally in Scylla.
The procedure is as follows:
1. Check if the user already exists in Scylla locally
2. If not, check if the user already has roles in LDAP
3. If some roles are assigned, auto-create the user in Scylla locally

Fixes #1919
Tests: manual (ldap + saslauthd auth)