Database Test: Add unit test that proves separation of concurrency semaphores
We include here a test to validate that indeed we created semaphore
separation for different service levels (reader concurrency semaphore).
The test is included separately because it includes additional code
changes for facilitating the test (i.e adding a spaceship default
operator in order to get automatic comparison).
The test validates that:
When a read query is ran in a specific service level, **only**
the stats for the concurrency semaphore associated with that particular
service level change and all other stats stay as before. The test also
validates that the correct memory distribution between semaphores holds
as new service levels (and as a result new semaphores) are being
created.
This proves at least the following:
1. The semaphores are different for different service levels.
2. There is isolation at least for the statistics that are measured by
   the semaphores themselves.
3. The memory distribution eventually holds.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>