Service Level Controller: Fix service level options when adding or updating a new service level
Up until now, the service level options that were stored in the database
were unchanged, even if some defaults had to be added in order to
continue (i.e decide how many shares to give to the different schedulers
groups). The reason for not having some options set is possible upgrade
from OSS where some of the options doesn't exist but are mandatory in
the enterprise version.
There can be a similar situation in the OSS but this hasn't occurred yet.
After this change the defaulted shares will be recorded and as a
consequence will be notified to the subscribers. This will allow them to
also become aware of the defaults that were used  and not only to the user
given values.
The unit test was changed accordingly.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>