test.py: Fail early when the AF_UNIX path is too long
There is a limitation on the socket path length for unix sockets
(see: `man 7 unix`) , when test.py is being run with such a tmpdir
that makes saslauthd socket path too long, the test will hang and
eventually fail after some retries and back offs.
It is pointless to keep trying so this change will make those test
fail early. In addition it will throw an exception that will
indicate the actual problem.

Refs: #1971

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>