test.py: Report test setup failures instead of just die
Up until now when one of the tests fails in the setup phase,
the whole execution of test.py dies on the spot. The reason was
lack of exception handling.
This change will print the exception and error details to the test
log and will report failure instead of just taking the whole run
down.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>