compaction/ICS: Optimize off-strategy compaction for node operations
Off-strategy compactions works by reshaping the maintenance set until
it's ready for integration into the main sstable set.

We can take advantage of the disjointness of the elements in the
maintenance set and compact them all together in a single round,
as the incremental reader selector will be able to open one element
at once, making the operation very memory efficient.
This also has the advantage of reducing write amplification as
we won't possibly have to recompact the same data in another
compaction round.

Fixes #1882.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20220324183638.78775-1-raphaelsc@scylladb.com>