compaction: ICS: Implement cleanup method for bucket awareness
Bucket awareness in cleanup was introduced in a69d98c3d0cba01.
All OSS strategies already implement it, now it's ICS turn to
have it.

The goal of bucket awareness is to reduce writeamp in cleanup,
therefore reducing operation time. Additionally, garbage collection
becomes more efficient as shadowed data can now be potentially
compacted with the data that shadows it, assuming they're on
the same tier.

The implementation for ICS is essentially the same as STCS one,
but changed to work with runs.

Fixes #2178.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>