kmip_host: Remove use of "cryptographic_parameters" in query
Refs #2205

The KMIP server in the issue above pukes when this seemingly harmless
attribute is present in a lookup. This is weird, but not a deal-breaker
since its usage is mainly just advisory anyway. It is by no means required
for finding/creating a key. So to adapt to least common denominator we
can just remove it.