ldap tests: Make slapd take reasonable amount of memory
As explained in #2249 , slapd takes a lot of memory due to the high
max open files limit in the testing environment (needed by scylla).
This patch implements the solution proposed to use prlimit with
reduced limits only on slapd so it will consume (a lot!!!) less
memory and will not fail on failure to allocate.

Tests: ldap unit tests in debug and dev mode.

Fixes #2249

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>

Closes #2250