scripts/pull_github_pr.sh: don't rely on environment variables
Environement variables can be easily leaked by various utilities that
collect environment information. Specifically, Fedora's bug reporting
ABRT collects environment variables and sends them in the report. It
offers to review the variables, and marks the report as private if a github
token is detected, but still better to avoid the issue in the first place.

This patch adds the ability to read the configuration from a file. We
use the same file that the `gh` utility uses, so authentication can
be done via `gh auth login`.

Environment variables are still supported and take precedence over the
configuration file.

Closes #2282