service_level_controller: avoid unexpected behavior after some races
Even though there is a gurad in place to protect against the same
node adding service_levels that don't have a backing scheduling group.
This doesn't protect us against several users adding service levels at
the same time. Once that happens, the behaviour is undefined.
This change aims to have a defined behaviour given that such race
occurred. The behaviour will be to only configure the first N service
levels where N is the number of available scheduling groups.
This will ensure that __enevntually__ all nodes will settle on the same
scheduling groups. A warning will be given each time the configuration
is polled to indicate the service levels that have been left out.

A unit test is added which validates that after a race the service
levels for different nodes is the same.

Tests: - Unit (Dev)
       - Manual test by manipulating the table to imitate the result
       of a race condition.
       - The new unit test, before and after the fix.

Refs: #2256
Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>