Workload Prioritization: protect against having two "deleted" scheduling groups with identical names.
After some sequences of create and delete service levels, the nodes
would assert on a duplicate label registration.
The reason is that scheduling groups of deleted service levels get
a specific and deterministic name so it can be identified in the
metrics. However, is some sequence of additions and deletions occur,
it can end up in two scheduling groups that are renamed to the same
name. (after the final deletion).
An example to such a sequence:
1. Start with service level "a" and "b" defined.
2. Remove service level "b"
3. Remove service level "a"
4. Create service level "a" again
5. Remove service level "a"

After step 5 the double registration assert will happen.

This change make's sure that this condition will not happen, by
making sure that if there is a deleted scheduling group for a recreated
service level it will be reused. This produces the invariant that at any
given moment, for a service level name X __exactly__ one of the following
holds:
1. It has no scheduling groups (A not yet created service level)
2. It has a scheduling group and the service level exists
3. It has a scheduling group in the deleted list and the service level
   does not exist.

A unit test was added to validate the fix.

Tests: - Unit Tests (dev)
       - Manual testing using the described above reproducer.
       - The new unit test before and after the fix

Refs: #2256
Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>