Messaging Service: generalize connection types per tenant
Commit e27af8a1bc466a87fcc8702f7532e62795821f8c shows that it is
reasonable to assume that the connection types per tenant will change
with versions, however, today it requires changes in multiple places in
the code, it is prone to coding mistakes and obfuscate this essential
information.
This change aims to keep the existing functionality while generalizing
the concept of a tenant connection type.

Also if fixes a bug in the isolation function where "forward" connection
types were assigned the default scheduling group since it wasn't
considered as a connection type prefix.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>