cql3: query_processor: make `get_migration_manager` private
After previous commits it's no longer used outside `query_processor`.
Also remove the `const` version - not needed for anything.

Use the getter instead of directly accessing `_mm` in `query_processor`
methods. Later, the getter will include a safety protection in case
`_mm` is missing (which should not happen when the user tries to execute
a schema change, but we'll want to stay safe).