reader_concurrency_semaphore_group: simplify and fix adjust()
The current implementation of adjust() is very complex: it tries to
avoid total memory consumption overshooting, when redistributing memory
between the semaphores. This can happen if semaphores which loose memory
are fully loaded. The algorithm drains these semaphore first to their
new (lower) memory limit and only then redistributes the memory among
the other semaphores. This algorithm is complex and is trying to avoid a
situation which can happen during normal operations. Semaphore memory
consumption can and does overshoot, they are not designed to avoid this.
Therefore in my opinion this complex algorithm is unwarranted and
instead of trying to fix the bugs it has, this patch just replaces it
with a very simple algorithm, just setting each semaphore to the new
memory value.

The current algorithm had two notable bugs, one benign and one serious:
* A few bytes of memory is leaked on each redistributation, due to
  off-by-one errors. The code seems to be aware of off-by-ones but
  apparently fails to compensate for it fully. This is benign, we are
  talking about a few bytes out of hundreds of thousands or millions.
* After a few cycles of updates, the memory assigned to semaphores is
  not what it should be according to their weights. In a recent issue
  (#2436) we've seen a semaphore without any memory units, unable to
  admit any reads. This issue is very serious as it can lead to denial
  of service.

Although the redistribution algorithm does have tests, they did not
discover the redistribution bugs. So this patch also adds a fuzzer test
which randomly adds/removes/updates scheduling groups, making sure
per-semaphore and total memory amounts are correct on each step.