raft topology: skip `left_token_ring` state during `removenode`
The "tell the node to shut down" RPC would fail every time in the
removenode path (since the node is dead), which is kind of awkward.

Besides, for removenode we don't really need the `left_token_ring`
state, we don't need to coordinate with the node - writes destined for
it are failing anyway (since it's dead) and we can ban the node
immediately.

Remove the node from group 0 while in `write_both_read_new` transition
state (even when we implement abort, in this state it's too late to
abort, we're committed to removing the node - so it's fine to remove it
from group 0 at this point).