skip left_token_ring state from removenode
The "tell the node to shut down" RPC would fail every time, which is
kind of awkward.

For removenode we don't really need the left_token_ring state, we don't
need to coordinate with the node - writes destined for it are failing
anyway (since it's dead) and we can ban the node immediately.

We can also remove it from group 0 in `write_both_read_new` transition
state (even when we implement abort, in this state it's too late to
abort, we're committed to removing the node).